requirements.txt

streamlit
openai
pillow
google-cloud-vision
